# Speed up SWS Track Notes