Fixes #1614

Calling GuidToTrack in for-loops made some operations painfully slow, e.g. "cfillion_Duplicate tracks (preserving SWS track notes).lua".

I made track-to-tracknotes lookups much faster by just comparing project pointers and guids, which is sufficient to globally identify any track.

Duplicating 256 tracks with notes:

Before:
GetSWSTrackNotes took 888 ms
Duplicate tracks took 41 ms
SetSWSTrackNotes took 4692 ms
The entire lua script took 5637 ms

Now:
GetSWSTrackNotes took: 1 ms
Duplicate tracks took: 53 ms
SetSWSTrackNotes took: 2 ms
The entire lua script took: 67 ms

It's barely even measureable in ms which made me worried that I screwed up something, but it seems to be working by my testing.